requirements.txt
Flask
Flask-PyMongo
bcrypt
python-dotenv
gunicorn